Backend-API/backend_api/requirements.txt
validate_email==1.3
opencv_python==4.5.2.52
Django==3.2.3
facenet==1.0.5
djangorestframework_api_key==2.0.0
drf_yasg==1.20.0
tensorflow==1.15.0
six==1.15.0
numpy==1.19.2
djangorestframework==3.12.4

tensorflow-gpu~=1.15.0
opencv-python~=4.5.2.52
pandas~=1.1.5
requests~=2.25.1
pdfkit~=0.6.1